requirements.txt
pandas==2.2.2
Pillow==9.0.1
